Fixed correctly
We were using only the first element of the pod. However, this is a
major change in interface, because load now returns an array with at
least one element, but array nonetheless. That's going to break
everything. Well, the two functions that use it, anyway.

This closes #3, and works towards perl6/Pod-To-HTML#55. Going back there now